V1/transcription-server/requirements.txt
nemo_toolkit[asr]
fastapi
uvicorn
soundfile
